Fix for issue #13
Change the return type of method QuerySorter.isSPARQL(String) from boolean to Query to reduce the number of calls for isSPARQL as parsing is expensive.
The fix is mainly handled by doing a recursion through the Execption block. Once a prettified error occurred, fix it, return to the same function. See QuerySorter.isSPARQL.

When tested with 2010-05-28.log the number of queries printed to queryset.log increases from 3797 to 4580 paring in mind the the only-complex-queries is set to true, which means that the increase may be a lower pound.